postcode distance in results.js - part 1
